code.sk
This skript will convert TEXT to binary and send it down a redstone line where it will be received by a receiver and be converted back into text. because minecraft updates 20 times a second, the upload and download speed cant be fast. i was working on a redstone alternative called fibreoptic cables but i realized it would still be bottle-necked by the compute speed of a Minecraft server. But it works nonthelesss.